backend/requirements.txt
Flask==2.3.2
python-dotenv==0.19.2
Authlib==0.15.5
google-cloud-logging
google-cloud-ndb
google-cloud-recaptcha-enterprise
google-cloud-secret-manager
absl-py